pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "fuzzywuzzy>=0.18.0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.2",
    "requests>=2.32.4",
    "telegram>=0.0.1",
]
